[data] split dask and modin tests
as dask tests will require running on python 3.12+
after upgrade; modin does not
